build: convert admin, blocks and tests/e2e to flat config
This retires the last four eslintrc files. No .eslintrc* or .eslintignore
remains anywhere in the repo.

admin and blocks both registered plugins the shared config already provides.
They must not: eslint-plugin-import has no ESLint v10 support and only works
through the fixupPluginRules-wrapped instance WordPress registers, and a second
copy of any plugin trips "Cannot redefine plugin" because flat config keys
plugins by object identity. So `plugins: ['import']`, `plugin:import/errors`,
`plugin:jest/recommended` and `plugin:@typescript-eslint/recommended` are gone,
and the rules they contributed are set directly on the inherited plugins. The
now-unused eslint-plugin-import, eslint-plugin-react, eslint-plugin-jest and
@typescript-eslint/* declarations are dropped from those packages.

blocks' `excludedFiles` becomes `ignores` alongside `files` in the same object,
and eslintrc's bare `*.ts` glob becomes `**/*.ts` - flat config matches full
paths, not basenames. `woocommerce/feature-flag: off` is dropped: the local
eslint-plugin-woocommerce ships no rules at all, its rules/ directory holding
only the requireindex barrel, so the rule it disabled no longer exists.

eslint-plugin-storybook moves back to ^0.12.0. I had bumped it to ^10.4.6 for
ESLint 10, but 9.x and 10.x peer-require storybook ^9/^10 while blocks runs 7.6.
0.12.0 peers eslint >=8 and ships flat/recommended.

tests/e2e has no package.json, so nothing ever linted it directly - it was
reached by eslintrc cascade from plugins/woocommerce. Flat config does not
cascade, so its rules move into that config, scoped to tests/e2e/**.
eslint-plugin-rulesdir has no flat equivalent, working by mutating a
module-level RULES_DIR, so no-raw-playwright-test-import is imported and
registered as a plugin object instead, and the dependency is dropped.

Verified: blocks applies its TypeScript override to block.tsx but not to
test/block.tsx, import/order reports through the wrapped plugin in both admin
and blocks, both lodash guards fire, and the e2e blocks subtree still reports
no-raw-playwright-test-import.

Part of #66078.
