--- plugins/woocommerce/client/admin/.eslintignore
@@ -1,15 +1,0 @@
-bin/*
-!bin/generate-docs
-!.eslintrc.js
-build
-build-module
-coverage
-languages
-node_modules
-vendor
-legacy
-tests/e2e
-build-types
-
-# These packages have their own eslint config and command
-api
