--- plugins/woocommerce/client/admin/.eslintrc.js
@@ -1,53 +1,0 @@
-module.exports = {
-	extends: [ 'plugin:@woocommerce/eslint-plugin/recommended' ],
-	plugins: [ 'import' ],
-	root: true,
-	overrides: [
-		{
-			files: [ 'client/**/*.js', 'client/**/*.jsx', 'client/**/*.tsx' ],
-			rules: {
-				'react/react-in-jsx-scope': 'off',
-			},
-		},
-	],
-	settings: {
-		'import/core-modules': [
-			'@woocommerce/admin-layout',
-			'@woocommerce/components',
-			'@woocommerce/customer-effort-score',
-			'@woocommerce/currency',
-			'@woocommerce/csv-export',
-			'@woocommerce/data',
-			'@woocommerce/date',
-			'@woocommerce/explat',
-			'@woocommerce/internal-js-tests',
-			'@woocommerce/navigation',
-			'@woocommerce/number',
-			'@woocommerce/onboarding',
-			'@woocommerce/product-editor',
-			'@woocommerce/settings',
-			'@woocommerce/tracks',
-			'@woocommerce/experimental',
-			'@wordpress/components',
-			'@wordpress/core-data',
-			'@wordpress/element',
-			'@wordpress/blocks',
-			'@wordpress/block-editor',
-			'@wordpress/block-library',
-			'@wordpress/notices',
-			'dompurify',
-			'@react-spring/web',
-			'react-router-dom',
-			'redux',
-			'xstate',
-			'xstate5',
-		],
-		'import/resolver': {
-			node: {},
-			webpack: {},
-			typescript: {
-				project: [ 'plugins/woocommerce/client/admin/tsconfig.json' ],
-			},
-		},
-	},
-};
