+++ plugins/woocommerce/client/admin/package.json
@@ -23,9 +23,9 @@
 		"lint": "pnpm --if-present '/^lint:lang:.*$/'",
 		"lint:fix": "pnpm --if-present '/^lint:fix:lang:.*$/'",
 		"lint:fix:lang:css": "stylelint '**/*.scss' --fix --ip 'storybook/wordpress'",
-		"lint:fix:lang:js": "pnpm lint:lang:js --fix --ext=js,ts,tsx",
+		"lint:fix:lang:js": "pnpm lint:lang:js --fix",
 		"lint:lang:css": "stylelint '**/*.scss' --cache --cache-location=node_modules/.cache/stylelint",
-		"lint:lang:js": "eslint ./client --ext=js,ts,tsx --cache --cache-location=node_modules/.cache/eslint",
+		"lint:lang:js": "eslint ./client --cache --cache-location=node_modules/.cache/eslint",
 		"lint:lang:types": "tsc --build --emitDeclarationOnly",
 		"postinstall": "node bin/xstate5-symlink.js",
 		"test:js": "jest --config client/jest.config.js",
@@ -159,6 +159,7 @@
 		"@typescript-eslint/eslint-plugin": "^8.63.0",
 		"@typescript-eslint/parser": "^8.63.0",
 		"@woocommerce/dependency-extraction-webpack-plugin": "workspace:*",
+		"@woocommerce/eslint-config": "workspace:*",
 		"@woocommerce/eslint-plugin": "workspace:*",
 		"@woocommerce/experimental-products-app": "workspace:*",
 		"@woocommerce/internal-build": "workspace:*",
@@ -187,8 +188,6 @@
 		"eslint": "^10.0.0",
 		"eslint-import-resolver-typescript": "^4.4.5",
 		"eslint-import-resolver-webpack": "^0.13.11",
-		"eslint-plugin-import": "^2.32.0",
-		"eslint-plugin-react": "^7.37.5",
 		"fs-extra": "11.1.1",
 		"jest": "29.5.x",
 		"jest-environment-jsdom": "29.5.x",
