--- plugins/woocommerce/client/blocks/.eslintignore
@@ -1,14 +1,0 @@
-!.eslintrc.js
-build
-build-module
-coverage
-languages
-node_modules
-vendor
-legacy
-reports
-tests/e2e-jest/specs/backend/__fixtures__
-tests/e2e-jest/specs/backend/__snapshots__
-
-storybook/dist
-assets/js/interactivity
