+++ plugins/woocommerce/client/blocks/package.json
@@ -60,9 +60,9 @@
 		"lint:ci": "pnpm run lint:js && pnpm run lint:css",
 		"lint:css": "stylelint '**/*.scss' --cache --cache-location=node_modules/.cache/stylelint",
 		"lint:css-fix": "stylelint '**/*.scss' --fix",
-		"lint:js": "WP_EXPERIMENTAL_MODULES=true wp-scripts lint-js --ext=js,ts,tsx --cache --cache-location=node_modules/.cache/eslint",
-		"lint:js:report": "pnpm run lint:js -- --output-file eslint_report.json --ext=js,ts,tsx --format json",
-		"lint:js-fix": "eslint assets/js --ext=js,jsx,ts,tsx --fix",
+		"lint:js": "WP_EXPERIMENTAL_MODULES=true wp-scripts lint-js --cache --cache-location=node_modules/.cache/eslint",
+		"lint:js:report": "pnpm run lint:js -- --output-file eslint_report.json --format json",
+		"lint:js-fix": "eslint assets/js --fix",
 		"lint:md:docs": "wp-scripts lint-md-docs",
 		"pre-commit": "lint-staged",
 		"reformat-files": "prettier --write \"**/*.{js,jsx,ts,tsx,css,scss}\"",
@@ -124,12 +124,11 @@
 		"@types/wordpress__block-editor": "catalog:wp-min",
 		"@types/wordpress__block-library": "catalog:wp-min",
 		"@types/wordpress__blocks": "catalog:wp-min",
-		"@typescript-eslint/eslint-plugin": "^8.63.0",
-		"@typescript-eslint/parser": "^8.63.0",
 		"@woocommerce/data": "workspace:*",
 		"@woocommerce/dependency-extraction-webpack-plugin": "workspace:*",
-		"@woocommerce/internal-build": "workspace:*",
+		"@woocommerce/eslint-config": "workspace:*",
 		"@woocommerce/eslint-plugin": "workspace:*",
+		"@woocommerce/internal-build": "workspace:*",
 		"@wordpress/api-fetch": "catalog:wp-min",
 		"@wordpress/babel-preset-default": "next",
 		"@wordpress/base-styles": "catalog:wp-bundled",
@@ -180,16 +179,13 @@
 		"dotenv": "^16.3.1",
 		"eslint-import-resolver-typescript": "^4.4.5",
 		"eslint-import-resolver-webpack": "^0.13.11",
-		"eslint-plugin-import": "^2.32.0",
-		"eslint-plugin-jest": "^29.15.4",
 		"eslint-plugin-playwright": "^2.10.5",
-		"eslint-plugin-rulesdir": "^0.2.2",
-		"eslint-plugin-storybook": "^10.4.6",
-		"eslint-plugin-woocommerce": "file:bin/eslint-plugin-woocommerce",
+		"eslint-plugin-storybook": "^0.12.0",
 		"eslint-plugin-you-dont-need-lodash-underscore": "^6.14.0",
 		"fs-extra": "11.1.1",
 		"gh-pages": "5.0.0",
 		"glob": "^10.3.10",
+		"globals": "^16.5.0",
 		"handlebars": "^4.7.8",
 		"ignore-loader": "0.1.x",
 		"jest": "29.5.x",
